+++ tests/creator/test_dataspace_creator.py
@@ -12,8 +12,8 @@ class TestDataspaceCreatorExample1:
     @pytest.fixture
     def dataspace_creator(self):
         creator = DataspaceCreator()
-        creator.add_dim("pressure.index", (0, 3), np.uint64)
-        creator.add_dim("temperature", (1, 8), np.uint64)
+        creator.add_shared_dim("pressure.index", (0, 3), np.uint64)
+        creator.add_shared_dim("temperature", (1, 8), np.uint64)
         creator.add_array_creator("A1", ("pressure.index",))
         filters = tiledb.FilterList(
             [
@@ -164,7 +164,7 @@ def test_create_array_no_array_error():
 
 def test_array_name_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (0, 3), np.int64)
+    creator.add_shared_dim("row", (0, 3), np.int64)
     creator.add_array_creator("array1", ("row",))
     with pytest.raises(ValueError):
         creator.add_array_creator("array1", ("row",))
@@ -172,7 +172,7 @@ def test_array_name_exists_error():
 
 def test_array_name_reserved_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (0, 3), np.int64)
+    creator.add_shared_dim("row", (0, 3), np.int64)
     with pytest.raises(ValueError):
         creator.add_array_creator(METADATA_ARRAY_NAME, ("row",))
 
@@ -185,7 +185,7 @@ def test_add_attr_no_array_error():
 
 def test_add_attr_name_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (0, 3), np.int64)
+    creator.add_shared_dim("row", (0, 3), np.int64)
     creator.add_array_creator("array1", ("row",))
     creator.add_array_creator("array2", ("row",))
     creator.add_attr_creator("attr1", "array1", np.float64)
@@ -195,7 +195,7 @@ def test_add_attr_name_exists_error():
 
 def test_add_attr_dim_name_in_array_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (0, 3), np.uint64)
+    creator.add_shared_dim("row", (0, 3), np.uint64)
     creator.add_array_creator("array1", ("row",))
     with pytest.raises(ValueError):
         creator.add_attr_creator("row", "array1", np.float64)
@@ -203,7 +203,7 @@ def test_add_attr_dim_name_in_array_exists_error():
 
 def test_add_attr_axis_data_coord_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (0, 3), np.int64)
+    creator.add_shared_dim("row", (0, 3), np.int64)
     creator.add_array_creator("array1", ("row",))
     creator.add_array_creator("array2", ("row",))
     creator.add_attr_creator("attr1", "array1", np.float64)
@@ -213,9 +213,9 @@ def test_add_attr_axis_data_coord_exists_error():
 
 def test_add_dim_name_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", (1, 4), np.uint64)
+    creator.add_shared_dim("row", (1, 4), np.uint64)
     with pytest.raises(ValueError):
-        creator.add_dim("row", (0, 3), np.int64)
+        creator.add_shared_dim("row", (0, 3), np.int64)
 
 
 def test_get_property_attr_key_error():
@@ -226,7 +226,7 @@ def test_get_property_attr_key_error():
 
 def test_remove_empty_array():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 10], np.int64)
+    creator.add_shared_dim("row", [0, 10], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.remove_array("A1")
     assert set(creator.array_names) == set()
@@ -234,7 +234,7 @@ def test_remove_empty_array():
 
 def test_remove_array_with_attrs():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 10], np.int64)
+    creator.add_shared_dim("row", [0, 10], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_array_creator("A2", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
@@ -247,7 +247,7 @@ def test_remove_array_with_attrs():
 
 def test_remove_renamed_array():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 10], np.int64)
+    creator.add_shared_dim("row", [0, 10], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_array_creator("A2", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
@@ -261,7 +261,7 @@ def test_remove_renamed_array():
 
 def test_remove_attr():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
     assert set(creator.attr_names) == {"x1"}
@@ -274,7 +274,7 @@ def test_remove_attr():
 
 def test_remove_attr_no_attr_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     with pytest.raises(KeyError):
         creator.remove_attr("x1")
@@ -282,7 +282,7 @@ def test_remove_attr_no_attr_error():
 
 def test_remove_dim():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int32)
+    creator.add_shared_dim("row", [0, 7], np.int32)
     assert set(creator.dim_names) == set(["row"])
     creator.remove_dim("row")
     assert set(creator.dim_names) == set()
@@ -290,7 +290,7 @@ def test_remove_dim():
 
 def test_remove_dim_axis_data():
     creator = DataspaceCreator()
-    creator.add_dim("row.data", [0.0, 100.0], np.float64)
+    creator.add_shared_dim("row.data", [0.0, 100.0], np.float64)
     assert set(creator.dim_names) == {"row.data"}
     creator.remove_dim("row.data")
     assert set(creator.dim_names) == set()
@@ -298,7 +298,7 @@ def test_remove_dim_axis_data():
 
 def test_remove_dim_in_use_error():
     creator = DataspaceCreator()
-    creator.add_dim("row.data", [0.0, 100.0], np.float64)
+    creator.add_shared_dim("row.data", [0.0, 100.0], np.float64)
     creator.add_array_creator("A1", ["row.data"], sparse=True)
     creator.add_array_creator("A2", ["row.data"], sparse=True)
     with pytest.raises(ValueError):
@@ -307,7 +307,7 @@ def test_remove_dim_in_use_error():
 
 def test_rename_array():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     assert set(creator.array_names) == {"A1"}
     creator.rename_array("A1", "B1")
@@ -316,8 +316,8 @@ def test_rename_array():
 
 def test_rename_array_with_attrs():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 3], np.int32)
-    creator.add_dim("col", [0, 3], np.int32)
+    creator.add_shared_dim("row", [0, 3], np.int32)
+    creator.add_shared_dim("col", [0, 3], np.int32)
     creator.add_array_creator("A1", ["row", "col"], sparse=True, tiles=[2, 2])
     creator.add_attr_creator("x1", "A1", np.float64)
     creator.add_attr_creator("x2", "A1", np.float64)
@@ -328,7 +328,7 @@ def test_rename_array_with_attrs():
 
 def test_rename_array_name_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_array_creator("B1", ["row"])
     with pytest.raises(ValueError):
@@ -337,7 +337,7 @@ def test_rename_array_name_exists_error():
 
 def test_rename_attr():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
     assert set(creator.attr_names) == {"x1"}
@@ -349,7 +349,7 @@ def test_rename_attr():
 
 def test_rename_attr_with_set_attr_properties():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
     assert set(creator.attr_names) == {"x1"}
@@ -361,7 +361,7 @@ def test_rename_attr_with_set_attr_properties():
 
 def test_rename_attr_name_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
     creator.add_attr_creator("y1", "A1", np.float64)
@@ -371,7 +371,7 @@ def test_rename_attr_name_exists_error():
 
 def test_rename_attr_dim_name_in_array_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("y1", (0, 4), np.int32)
+    creator.add_shared_dim("y1", (0, 4), np.int32)
     creator.add_array_creator("A1", ["y1"])
     creator.add_attr_creator("x1", "A1", np.float64)
     with pytest.raises(ValueError):
@@ -380,7 +380,7 @@ def test_rename_attr_dim_name_in_array_exists_error():
 
 def test_rename_dim_not_used():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 3], np.int32)
+    creator.add_shared_dim("row", [0, 3], np.int32)
     assert set(creator.dim_names) == {"row"}
     creator.rename_dim("row", "col")
     assert set(creator.dim_names) == {"col"}
@@ -388,7 +388,7 @@ def test_rename_dim_not_used():
 
 def test_rename_dim_dataspace_axis():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 3], np.float64)
+    creator.add_shared_dim("row", [0, 3], np.float64)
     assert set(creator.dim_names) == {"row"}
     creator.rename_dim("row", "col")
     assert set(creator.dim_names) == {"col"}
@@ -396,7 +396,7 @@ def test_rename_dim_dataspace_axis():
 
 def test_rename_dim_used():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     assert set(creator.dim_names) == {"row"}
     creator.rename_dim("row", "unit")
@@ -414,23 +414,23 @@ def test_array_no_dims_error():
 
 def test_rename_dim_name_exists_in_dataspace_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 3], np.int32)
-    creator.add_dim("col", [0, 7], np.int32)
+    creator.add_shared_dim("row", [0, 3], np.int32)
+    creator.add_shared_dim("col", [0, 7], np.int32)
     with pytest.raises(NotImplementedError):
         creator.rename_dim("col", "row")
 
 
 def test_rename_dim_no_merge_error():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 3], np.int32)
-    creator.add_dim("col", [0, 3], np.int32)
+    creator.add_shared_dim("row", [0, 3], np.int32)
+    creator.add_shared_dim("col", [0, 3], np.int32)
     with pytest.raises(NotImplementedError):
         creator.rename_dim("col", "row")
 
 
 def test_rename_dim_attr_name__in_array_exists_error():
     creator = DataspaceCreator()
-    creator.add_dim("y1", (0, 4), np.int32)
+    creator.add_shared_dim("y1", (0, 4), np.int32)
     creator.add_array_creator("A1", ["y1"])
     creator.add_attr_creator("x1", "A1", np.float64)
     with pytest.raises(ValueError):
@@ -439,7 +439,7 @@ def test_rename_dim_attr_name__in_array_exists_error():
 
 def test_set_attr_property():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("A1", ["row"])
     creator.add_attr_creator("x1", "A1", np.float64)
     creator.set_attr_properties("x1", fill=-1)
@@ -450,7 +450,7 @@ def test_set_attr_property():
 
 def test_set_dim_dtype():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.set_dim_properties("row", dtype=np.float64)
     dtype = creator.get_dim_property("row", "dtype")
     assert dtype == np.dtype(np.float64)
@@ -458,7 +458,7 @@ def test_set_dim_dtype():
 
 def test_set_dim_domain():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.set_dim_properties("row", domain=(1, 8))
     domain = creator.get_dim_property("row", "domain")
     assert domain == (1, 8)
@@ -466,7 +466,7 @@ def test_set_dim_domain():
 
 def test_set_dim_name():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.set_dim_properties("row", name="column")
     name = tuple(creator.dim_names)[0]
     assert name == "column"
@@ -474,7 +474,7 @@ def test_set_dim_name():
 
 def test_set_dim_dtype_with_array_check():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, 7], np.int64)
+    creator.add_shared_dim("row", [0, 7], np.int64)
     creator.add_array_creator("array1", ("row",))
     creator.set_dim_properties("row", dtype=np.uint32)
     dtype = creator.get_dim_property("row", "dtype")
@@ -497,7 +497,7 @@ def test_dataspace_creator_name():
 
 def test_to_schema_bad_array():
     creator = DataspaceCreator()
-    creator.add_dim("row", [0, -1], np.int64)
+    creator.add_shared_dim("row", [0, -1], np.int64)
     creator.add_array_creator("A1", ("row",))
     creator.add_attr_creator("x1", "A1", dtype=np.float64)
     with pytest.raises(RuntimeError):